made all options in endpoint object in endpoints arrat accepts undefined emoved jest and tests
