+++ lib/index.js
@@ -1,44 +1,42 @@
 const { RawSource } = require('webpack-sources');
 const fetch = require('node-fetch');
-
-const waitFor = (ms) => new Promise(r => setTimeout(r, ms))
-
-const asyncForEach = async (array, callback) => {
-  for (let index = 0; index < array.length; index++) {
-    try {
-      await callback(array[index], index, array);
-    } catch (err) {
-      console.log(err);
-    }
-}
+const chalk = require('chalk');
+const { waitFor, asyncForEach } = require('./utils');
+const log = console.log;
 
 module.exports = class FetchJsonWebpackPlugin {
   constructor(options) {
     this.options = options || {};
     this.options.endpoints = options.endpoints || null;
-    this.options.hash = options.hash || false;
     this.createJson = this.createJson.bind(this);
   }
+  
   apply(compiler) {
     compiler.hooks.emit.tapAsync(this.constructor.name, this.createJson);
   }
-
+  
   createJson(compilation, cb) {
     
-    const returnHash = (filename) =>
-      this.options.hash
-        ? `${filename}.${compilation.hash}.json`
-        : `${filename}.json`;
+    const returnFileName = (filename, i) =>
+      filename === undefined
+        ? `data-${i + 1}`
+        : filename;
+
+    const returnHash = (name, hash) =>
+      (hash === undefined || hash === false)
+        ? `${name}.json`
+        : `${name}.${compilation.hash}.json`;
     
     const request = async () => {
       try {
-        await asyncForEach(this.options.endpoint, async (endpoint) => {
+        await asyncForEach(this.options.endpoints, async (endpoint, i) => {
           await waitFor(50);
           let data = {};
           const response = await fetch(endpoint.url);
           data = await response.json();
-          compilation.assets[returnHash(endpoint.filename)] = new RawSource(JSON.stringify(data));
-          console.log(`Successfully compiled ${endpoint.filename} from data endpoint!`);
+          const file = returnHash(returnFileName(endpoint.filename, i), endpoint.hash);
+          compilation.assets[file] = new RawSource(JSON.stringify(data));
+          log(chalk.yellow.bgRed.bold(`Successfully compiled endpoint file ${i + 1} as ${file} from data endpoint!`));
         });
         cb();
         console.log('Done');
@@ -47,6 +45,7 @@ module.exports = class FetchJsonWebpackPlugin {
         cb();
       }
     };
+    
     request();
   }
 };
