+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fetch-json-webpack-plugin",
-  "version": "1.0.2",
+  "version": "2.0.0",
   "description": "Webpack plugin to generate JSON assets from data api endpoints.",
   "author": "Dain Blodorn Kim",
   "license": "MIT",
@@ -31,16 +31,13 @@
   },
   "homepage": "https://github.com/dblodorn/fetch-json-webpack-plugin#readme",
   "devDependencies": {
+    "chalk": "^3.0.0",
     "clean-webpack-plugin": "^1.0.0",
     "eslint": "^5.9.0",
-    "jest": "^23.6.0",
+    "node-fetch": "^2.3.0",
     "webpack": "^4.26.1",
     "webpack-cli": "^3.1.2",
-    "webpack-merge": "^4.1.4"
-  },
-  "dependencies": {
-    "chalk": "^3.0.0",
-    "node-fetch": "^2.3.0",
+    "webpack-merge": "^4.1.4",
     "webpack-sources": "^1.3.0"
   }
 }
